event-generator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vehicle-event-generator"
version = "1.0.0"
description = "Vehicle location event generator for Azure Event Hubs"
authors = [
    {name = "Azure Dev Summit", email = "demo@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "azure-eventhub>=5.11.1",
    "azure-identity>=1.15.0",
    "cloudevents>=1.10.1",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
vehicle-generator = "vehicle_generator.cli:main"

[project.urls]
Homepage = "https://github.com/your-repo/eventhubs-demo"
Repository = "https://github.com/your-repo/eventhubs-demo.git"
Issues = "https://github.com/your-repo/eventhubs-demo/issues"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]